requirements.txt
python-telegram-bot==13.8.1
python-decouple==3.4
boto3==1.18.32
pytz==2020.5